Generalize the script
This commit removes the SW references and trims a lot of the fat. In
this state it can be shared as an auto-activation script.

One precedent here is that we don't commit a version bump. Instead, the
version number is updated as soon as it is warranted with the suffix
~dev. When a version is stable and tested, it is pinned with its own
commit that removes the suffix. Commits that pin also get a tag. If we
go with a release branch, these can be cherry-picked after the fact.