Merge: Join Team
This change removes the test command entirely, and reduces the set of
possible teams down to just red, green, blue, and yellow. 

It adds the "--join-team" command which exclusively adds a coloured badge
to selected tokens. Because teams are tied to Auto Activation, the command
also resets the activation status by removing the associated badge.